NP-2569: Changed validation from double to integer
Added more error logging

+++ src/main/java/no/unit/nva/cristin/projects/FetchCristinProjects.java
@@ -104,7 +104,7 @@ public class FetchCristinProjects extends CristinHandler<Void, ProjectsWrapper> 
 
     private boolean isNumeric(String str) {
         try {
-            Double.parseDouble(str);
+            Integer.parseInt(str);
             return true;
         } catch (NumberFormatException e) {
             return false;
